api/requirements.txt
# RecyClique API — Python 3.12
# Story 1.2+2.1: FastAPI, health, DB, sites CRUD
# Story 3.1: JWT auth (python-jose, passlib)

fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic-settings>=2.6.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.14.0
redis>=5.0.0
httpx>=0.27.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pytest>=8.0.0
